+++ gen/com/intellij/idea/plugin/hybris/impex/psi/ImpexTypes.java
@@ -95,6 +95,7 @@ public interface ImpexTypes {
   IElementType FIELD_VALUE_IGNORE = new ImpexTokenType("FIELD_VALUE_IGNORE");
   IElementType FIELD_VALUE_JAR_PREFIX = new ImpexTokenType("FIELD_VALUE_JAR_PREFIX");
   IElementType FIELD_VALUE_NULL = new ImpexTokenType("FIELD_VALUE_NULL");
+  IElementType FIELD_VALUE_PASSWORD_ENCODING_PREFIX = new ImpexTokenType("FIELD_VALUE_PASSWORD_ENCODING_PREFIX");
   IElementType FIELD_VALUE_SEPARATOR = new ImpexTokenType("FIELD_VALUE_SEPARATOR");
   IElementType FIELD_VALUE_ZIP_PREFIX = new ImpexTokenType("FIELD_VALUE_ZIP_PREFIX");
   IElementType FUNCTION = new ImpexTokenType("FUNCTION");